Changed blogging/blog/models.py
I set the database layout for the blog application by following the "Creating model" section of 2nd part of the official django tutorial (source:https://docs.djangoproject.com/en/1.9/intro/tutorial02/). The main field of the database are the post text and the comment text. There can be other fields to add (unlike), so I let anybody else to complete that.